requirements.txt
colorlog
pynacl